Add files via upload
if crashing and lagging your pc and blocking mouse and keyboard input counts as a "malicious program" DO NOT open any batch file with "Test" in the name. The only .bat that will remove lightspeed filter agent from your computer is the one with it in the name. 
This is a safe folder, any suspicious feelings can be shooed away by a simple virus scan, or opening the files with Dnspy to see its true contents.
Enjoy your anti filter, and FUCK LIGHTSPEED FILTER AGENT BABYY